+++ packages/snaps-controllers/src/snaps/SnapController.ts
@@ -161,15 +161,15 @@ import {
 import type { SnapLocation } from './location';
 import { detectSnapLocation } from './location';
 import type {
-  GetMetadata,
-  GetResult,
-  ResolveVersion,
-  SnapsRegistryInfo,
-  SnapsRegistryRequest,
-  SnapsRegistryStateChangeEvent,
-  Update,
+  SnapRegistryControllerGetAction,
+  SnapRegistryControllerGetMetadataAction,
+  SnapRegistryControllerResolveVersionAction,
+  SnapRegistryControllerRequestUpdateAction,
+  SnapRegistryInfo,
+  SnapRegistryRequest,
+  SnapRegistryControllerStateChangeEvent,
 } from './registry';
-import { SnapsRegistryStatus } from './registry';
+import { SnapRegistryStatus } from './registry';
 import { getRunnableSnaps } from './selectors';
 import type { SnapControllerMethodActions } from './SnapController-method-action-types';
 import { Timer } from './Timer';
@@ -530,10 +530,10 @@ export type AllowedActions =
   | ExecutionServiceTerminateSnapAction
   | UpdateCaveat
   | ApprovalControllerUpdateRequestStateAction
-  | GetResult
-  | GetMetadata
-  | Update
-  | ResolveVersion
+  | SnapRegistryControllerGetAction
+  | SnapRegistryControllerGetMetadataAction
+  | SnapRegistryControllerResolveVersionAction
+  | SnapRegistryControllerRequestUpdateAction
   | SnapInterfaceControllerCreateInterfaceAction
   | SnapInterfaceControllerGetInterfaceAction
   | SnapInterfaceControllerSetInterfaceDisplayedAction
@@ -547,7 +547,7 @@ export type AllowedEvents =
   | SnapControllerSnapInstalledEvent
   | SnapControllerSnapUpdatedEvent
   | KeyringControllerLockEvent
-  | SnapsRegistryStateChangeEvent;
+  | SnapRegistryControllerStateChangeEvent;
 
 export type SnapControllerMessenger = Messenger<
   typeof controllerName,
@@ -975,7 +975,7 @@ export class SnapController extends BaseController<
     );
 
     this.messenger.subscribe(
-      'SnapsRegistry:stateChange',
+      'SnapRegistryController:stateChange',
       () => {
         this.#handleRegistryUpdate().catch((error) => {
           logError(
@@ -998,7 +998,7 @@ export class SnapController extends BaseController<
     this.#trackSnapExport = throttleTracking(
       (snapId: SnapId, handler: string, success: boolean, origin: string) => {
         const snapMetadata = this.messenger.call(
-          'SnapsRegistry:getMetadata',
+          'SnapRegistryController:getMetadata',
           snapId,
         );
         this.#trackEvent({
@@ -1452,7 +1452,7 @@ export class SnapController extends BaseController<
    */
   async updateRegistry(): Promise<void> {
     await this.#ensureCanUsePlatform();
-    await this.messenger.call('SnapsRegistry:update');
+    await this.messenger.call('SnapRegistryController:requestUpdate');
   }
 
   /**
@@ -1464,8 +1464,8 @@ export class SnapController extends BaseController<
    */
   async #handleRegistryUpdate() {
     const blockedSnaps = await this.messenger.call(
-      'SnapsRegistry:get',
-      Object.values(this.state.snaps).reduce<SnapsRegistryRequest>(
+      'SnapRegistryController:get',
+      Object.values(this.state.snaps).reduce<SnapRegistryRequest>(
         (blockListArg, snap) => {
           blockListArg[snap.id] = {
             version: snap.version,
@@ -1479,7 +1479,7 @@ export class SnapController extends BaseController<
 
     await Promise.all(
       Object.entries(blockedSnaps).map(async ([snapId, { status, reason }]) => {
-        if (status === SnapsRegistryStatus.Blocked) {
+        if (status === SnapRegistryStatus.Blocked) {
           return this.#blockSnap(snapId as SnapId, reason);
         }
 
@@ -1586,17 +1586,17 @@ export class SnapController extends BaseController<
     {
       platformVersion,
       ...snapInfo
-    }: SnapsRegistryInfo & {
+    }: SnapRegistryInfo & {
       permissions: SnapPermissions;
       platformVersion: string | undefined;
     },
   ) {
-    const results = await this.messenger.call('SnapsRegistry:get', {
+    const results = await this.messenger.call('SnapRegistryController:get', {
       [snapId]: snapInfo,
     });
 
     const result = results[snapId];
-    if (result.status === SnapsRegistryStatus.Blocked) {
+    if (result.status === SnapRegistryStatus.Blocked) {
       throw new Error(
         `Cannot install version "${
           snapInfo.version
@@ -1613,11 +1613,11 @@ export class SnapController extends BaseController<
     if (
       this.#featureFlags.requireAllowlist &&
       isAllowlistingRequired &&
-      result.status !== SnapsRegistryStatus.Verified
+      result.status !== SnapRegistryStatus.Verified
     ) {
       throw new Error(
         `Cannot install version "${snapInfo.version}" of snap "${snapId}": ${
-          result.status === SnapsRegistryStatus.Unavailable
+          result.status === SnapRegistryStatus.Unavailable
             ? 'The registry is temporarily unavailable.'
             : 'The snap is not on the allowlist.'
         }`,
@@ -3107,7 +3107,7 @@ export class SnapController extends BaseController<
     versionRange: SemVerRange,
   ): Promise<SemVerRange> {
     return await this.messenger.call(
-      'SnapsRegistry:resolveVersion',
+      'SnapRegistryController:resolveVersion',
       snapId,
       versionRange,
     );